move files to src/ and example/ directory.
