+++ example/example-timecraftjs.html
@@ -3,9 +3,9 @@
 	<head>
 		<meta charset="utf-8">
 		<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-		<script type="text/javascript" src="./spice.js"></script>
-		<script type="text/javascript" src="./timecraft.js"></script>
-		<script type="text/javascript" src="./cspice.js"></script>
+		<script type="text/javascript" src="../src/spice.js"></script>
+		<script type="text/javascript" src="../timecraft.js"></script>
+		<script type="text/javascript" src="../src/cspice.js"></script>
 	</head>
 	<body>
 		<script>
